+++ lib/Sema/CSBindings.cpp
@@ -20,6 +20,7 @@
 #include "swift/Basic/Assertions.h"
 #include "swift/Sema/ConstraintGraph.h"
 #include "swift/Sema/ConstraintSystem.h"
+#include "swift/Sema/Subtyping.h"
 #include "swift/Sema/TypeVariableType.h"
 #include "llvm/ADT/SetVector.h"
 #include "llvm/Support/Debug.h"
@@ -1488,80 +1489,6 @@ bool swift::constraints::inference::checkTypeOfBinding(
 
   // Okay, allow the binding.
   return true;
-}
-
-ConversionBehavior
-swift::constraints::inference::getConversionBehavior(Type type) {
-  if (type->is<StructType>()) {
-    if (type->isAnyHashable())
-      return ConversionBehavior::AnyHashable;
-    else if (type->isDouble() || type->isCGFloat())
-      return ConversionBehavior::Double;
-    else if (type->getAnyPointerElementType())
-      return ConversionBehavior::Pointer;
-
-    return ConversionBehavior::None;
-  }
-
-  if (auto *structTy = type->getAs<BoundGenericStructType>()) {
-    if (auto eltTy = structTy->getArrayElementType()) {
-      return ConversionBehavior::Array;
-    } else if (ConstraintSystem::isDictionaryType(structTy)) {
-      return ConversionBehavior::Dictionary;
-    } else if (ConstraintSystem::isSetType(structTy)) {
-      return ConversionBehavior::Set;
-    } else if (type->getAnyPointerElementType()) {
-      return ConversionBehavior::Pointer;
-    }
-
-    return ConversionBehavior::None;
-  }
-
-  if (auto *enumTy = type->getAs<BoundGenericEnumType>()) {
-    if (enumTy->getOptionalObjectType())
-      return ConversionBehavior::Optional;
-
-    return ConversionBehavior::None;
-  }
-
-  if (type->is<ClassType>() || type->is<BoundGenericClassType>())
-    return ConversionBehavior::Class;
-
-  if (type->is<EnumType>() ||
-      type->is<BuiltinType>() || type->is<ArchetypeType>())
-    return ConversionBehavior::None;
-
-  if (type->is<FunctionType>() || type->is<MetatypeType>())
-    return ConversionBehavior::Structural;
-
-  return ConversionBehavior::Unknown;
-}
-
-/// Check whether there exists a type that could be implicitly converted
-/// to a given type i.e. is the given type is Double or Optional<..> this
-/// function is going to return true because CGFloat could be converted
-/// to a Double and non-optional value could be injected into an optional.
-bool swift::constraints::inference::hasConversions(Type type) {
-  switch (getConversionBehavior(type)) {
-  case ConversionBehavior::None:
-    return false;
-  case ConversionBehavior::Array:
-    return hasConversions(type->getArrayElementType());
-  case ConversionBehavior::Dictionary: {
-    auto pair = ConstraintSystem::isDictionaryType(type);
-    return hasConversions(pair->first) || hasConversions(pair->second);
-  }
-  case ConversionBehavior::Set:
-    return hasConversions(*ConstraintSystem::isSetType(type));
-  case ConversionBehavior::Class:
-  case ConversionBehavior::AnyHashable:
-  case ConversionBehavior::Double:
-  case ConversionBehavior::Pointer:
-  case ConversionBehavior::Optional:
-  case ConversionBehavior::Structural:
-  case ConversionBehavior::Unknown:
-    return true;
-  }
 }
 
 bool BindingSet::isViable(PotentialBinding &binding) {
