+++ lib/Sema/CSTrail.cpp
@@ -27,6 +27,7 @@
 
 using namespace swift;
 using namespace constraints;
+using namespace inference;
 
 #define DEBUG_TYPE "SolverTrail"
 
@@ -324,7 +325,7 @@ SolverTrail::Change::RetiredConstraint(ConstraintList::iterator where,
 
 SolverTrail::Change
 SolverTrail::Change::AddedBinding(TypeVariableType *typeVar,
-                                  inference::PotentialBinding binding) {
+                                  PotentialBinding binding) {
   Change result;
   result.Kind = ChangeKind::AddedBinding;
   result.Binding.TypeVar = typeVar;
@@ -338,7 +339,7 @@ SolverTrail::Change::AddedBinding(TypeVariableType *typeVar,
 
 SolverTrail::Change
 SolverTrail::Change::RetractedBinding(TypeVariableType *typeVar,
-                                      inference::PotentialBinding binding) {
+                                      PotentialBinding binding) {
   Change result;
   result.Kind = ChangeKind::RetractedBinding;
   result.Binding.TypeVar = typeVar;
